refactor(sdk/history): align file layout with memory-refactor RFC
Strict alignment with §4.1 of docs/memory-refactor.md (Scheme A):

- Rename interface Memory→History (matches package name; eliminates
  collision with recall.Memory)
- File renames:
  - memory.go → history.go (interface + Budget only)
  - lossless.go → compactor.go (also absorbs summary_dag.go,
    summary_index.go, factory.go's NewCompacted/Options — all
    summarization concerns now live in one file)
  - file_store.go + InMemoryStore → store.go (one file for the
    persistence interface and shipped impls)
  - counter.go → token.go (matches RFC label "TokenCounter")
  - summary_prompt.go → compactor_prompt.go
- Internal compacted struct → compactor (matches file name)
- Add doc.go (package overview the RFC explicitly calls out)
- Drop factory.go entirely; NewBuffer lives next to buffer in
  buffer.go, NewCompacted next to compactor in compactor.go
- Test files renamed to mirror sources

Public API change: history.Memory → history.History. Single caller
(examples/chatbot-with-recall) updated. No production deployments
exist yet; no compat shim added.

verified: make ci + GOWORK=off bench + example demo all green.
Made-with: Cursor
