+++ sdk/history/compactor_factory_test.go
