+++ sdk/history/compactor_index_test.go
