+++ sdk/history/store.go
@@ -7,6 +7,7 @@ import (
 	"encoding/json"
 	"fmt"
 	"sync"
+	"time"
 
 	"github.com/GizClaw/flowcraft/sdk/model"
 	"github.com/GizClaw/flowcraft/sdk/telemetry"
@@ -14,6 +15,183 @@ import (
 
 	otellog "go.opentelemetry.io/otel/log"
 )
+
+// Store is the persistence-layer interface for short-term message storage.
+type Store interface {
+	GetMessages(ctx context.Context, conversationID string) ([]model.Message, error)
+	SaveMessages(ctx context.Context, conversationID string, messages []model.Message) error
+	DeleteMessages(ctx context.Context, conversationID string) error
+}
+
+// RecentReader is an optional interface for stores that can efficiently
+// read only the most recent N messages for a conversation.
+type RecentReader interface {
+	GetRecentMessages(ctx context.Context, conversationID string, limit int) ([]model.Message, error)
+}
+
+// MessageAppender is an optional interface for stores that can append
+// only the newly generated messages without reloading the full history.
+type MessageAppender interface {
+	AppendMessages(ctx context.Context, conversationID string, messages []model.Message) error
+}
+
+// RangeReader is an optional interface for stores that support
+// reading a subset of messages by sequence range.
+type RangeReader interface {
+	GetMessageRange(ctx context.Context, conversationID string, start, end int) ([]model.Message, error)
+}
+
+// SummaryCacheStore is an optional interface that Store implementations
+// can satisfy to persist summary caches alongside messages.
+type SummaryCacheStore interface {
+	GetSummary(ctx context.Context, conversationID string) (summary string, msgCount int, err error)
+	SaveSummary(ctx context.Context, conversationID, summary string, msgCount int) error
+}
+
+const (
+	defaultMaxConversations = 10000
+	defaultConversationTTL  = 24 * time.Hour
+)
+
+// InMemoryStoreOption configures an InMemoryStore.
+type InMemoryStoreOption func(*InMemoryStore)
+
+// WithMaxConversations sets the upper bound on stored conversations.
+// When exceeded, the least-recently-accessed conversation is evicted.
+func WithMaxConversations(n int) InMemoryStoreOption {
+	return func(s *InMemoryStore) {
+		if n > 0 {
+			s.maxConversations = n
+		}
+	}
+}
+
+// WithTTL sets how long an idle conversation is kept before eviction.
+func WithTTL(d time.Duration) InMemoryStoreOption {
+	return func(s *InMemoryStore) {
+		if d > 0 {
+			s.ttl = d
+		}
+	}
+}
+
+type memEntry struct {
+	messages   []model.Message
+	lastAccess time.Time
+}
+
+// InMemoryStore is a simple in-memory message store for development and testing.
+// It supports a maximum conversation count and TTL-based eviction.
+type InMemoryStore struct {
+	mu               sync.RWMutex
+	data             map[string]*memEntry
+	maxConversations int
+	ttl              time.Duration
+	done             chan struct{}
+}
+
+// NewInMemoryStore creates a new in-memory message store with optional limits.
+func NewInMemoryStore(opts ...InMemoryStoreOption) *InMemoryStore {
+	s := &InMemoryStore{
+		data:             make(map[string]*memEntry),
+		maxConversations: defaultMaxConversations,
+		ttl:              defaultConversationTTL,
+		done:             make(chan struct{}),
+	}
+	for _, opt := range opts {
+		opt(s)
+	}
+	go s.cleanupLoop()
+	return s
+}
+
+// Close stops the background cleanup goroutine.
+func (s *InMemoryStore) Close() {
+	select {
+	case <-s.done:
+	default:
+		close(s.done)
+	}
+}
+
+func (s *InMemoryStore) cleanupLoop() {
+	ticker := time.NewTicker(s.ttl / 2)
+	defer ticker.Stop()
+	for {
+		select {
+		case <-s.done:
+			return
+		case <-ticker.C:
+			s.evictExpired()
+		}
+	}
+}
+
+func (s *InMemoryStore) evictExpired() {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+	cutoff := time.Now().Add(-s.ttl)
+	for id, entry := range s.data {
+		if entry.lastAccess.Before(cutoff) {
+			delete(s.data, id)
+		}
+	}
+}
+
+// evictOldest removes the least-recently-accessed entry. Caller must hold s.mu.
+func (s *InMemoryStore) evictOldest() {
+	var oldestID string
+	var oldestTime time.Time
+	for id, entry := range s.data {
+		if oldestID == "" || entry.lastAccess.Before(oldestTime) {
+			oldestID = id
+			oldestTime = entry.lastAccess
+		}
+	}
+	if oldestID != "" {
+		delete(s.data, oldestID)
+	}
+}
+
+func (s *InMemoryStore) GetMessages(_ context.Context, conversationID string) ([]model.Message, error) {
+	s.mu.Lock()
+	entry := s.data[conversationID]
+	if entry == nil {
+		s.mu.Unlock()
+		return nil, nil
+	}
+	entry.lastAccess = time.Now()
+	cp := make([]model.Message, len(entry.messages))
+	copy(cp, entry.messages)
+	s.mu.Unlock()
+	return cp, nil
+}
+
+func (s *InMemoryStore) SaveMessages(_ context.Context, conversationID string, messages []model.Message) error {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+	cp := make([]model.Message, len(messages))
+	copy(cp, messages)
+	if _, exists := s.data[conversationID]; !exists && len(s.data) >= s.maxConversations {
+		s.evictOldest()
+	}
+	s.data[conversationID] = &memEntry{messages: cp, lastAccess: time.Now()}
+	return nil
+}
+
+func (s *InMemoryStore) DeleteMessages(_ context.Context, conversationID string) error {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+	delete(s.data, conversationID)
+	return nil
+}
+
+// Len returns the number of stored conversations (useful for testing/monitoring).
+func (s *InMemoryStore) Len() int {
+	s.mu.RLock()
+	defer s.mu.RUnlock()
+	return len(s.data)
+}
 
 // FileStore is a Workspace-backed Store that persists messages as JSONL files.
 // Each conversation is stored at {prefix}/{conversationID}/messages.jsonl with
